pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "stoi"
version = "2.0.0"
description = "STOI — Shit Token On Investment：量化 AI 编程工具的 Token 效率，定位浪费根因"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "KevinYoung-Kw"},
]
keywords = ["claude", "token-efficiency", "cli", "anthropic", "prompt-caching"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Quality Assurance",
    "Topic :: Utilities",
]
requires-python = ">=3.9"
dependencies = [
    "aiohttp>=3.9.0",
    "textual>=0.50.0",
    "rich>=13.0.0",
    "tavily-python>=0.5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "black>=23.0",
    "ruff>=0.1.0",
]

[project.scripts]
stoi = "stoi.stoi:main"

[project.urls]
Homepage = "https://github.com/KevinYoung-Kw/stoi"
Repository = "https://github.com/KevinYoung-Kw/stoi"
Issues = "https://github.com/KevinYoung-Kw/stoi/issues"

[tool.setuptools]
packages = ["stoi"]
include-package-data = true

[tool.setuptools.package-data]
stoi = ["stoi_knowledge/*.md", "stoi_knowledge/**/*.md"]

[tool.black]
line-length = 100
target-version = ['py39']

[tool.ruff]
line-length = 100
select = ["E", "F", "W", "I"]
